+++ SampleProjects/TestSomething/src/excludeThis/exclude-this.cpp
@@ -1,6 +1,0 @@
-This file intentionally contains syntactically incorrect code
-to break unit test compilation.  If arduino_ci is working
-properly, it should exclude this file (as per .arduino-ci.yml
-configuration) and unit test compilation should succeed.
-
-~!@#$%^&*()