+++ spec/cpp_library_spec.rb
@@ -14,58 +14,79 @@ RSpec.describe ArduinoCI::CppLibrary do
   next if skip_ruby_tests
 
   answers = {
-    "DoSomething": {
+    DoSomething: {
       one_five: false,
       cpp_files: [Pathname.new("DoSomething") + "do-something.cpp"],
+      cpp_files_libraries: [],
       header_dirs: [Pathname.new("DoSomething")],
+      arduino_library_src_dirs: [],
       test_files: [
         "DoSomething/test/good-null.cpp",
         "DoSomething/test/good-library.cpp",
         "DoSomething/test/bad-null.cpp",
       ].map { |f| Pathname.new(f) }
     },
-    "OnePointOhDummy": {
+    OnePointOhDummy: {
       one_five: false,
       cpp_files: [
         "OnePointOhDummy/YesBase.cpp",
         "OnePointOhDummy/utility/YesUtil.cpp",
       ].map { |f| Pathname.new(f) },
+      cpp_files_libraries: [],
       header_dirs: [
         "OnePointOhDummy",
         "OnePointOhDummy/utility"
       ].map { |f| Pathname.new(f) },
+      arduino_library_src_dirs: [],
       test_files: []
     },
-    "OnePointFiveMalformed": {
+    OnePointFiveMalformed: {
       one_five: false,
       cpp_files: [
         "OnePointFiveMalformed/YesBase.cpp",
         "OnePointFiveMalformed/utility/YesUtil.cpp",
       ].map { |f| Pathname.new(f) },
+      cpp_files_libraries: [],
       header_dirs: [
         "OnePointFiveMalformed",
         "OnePointFiveMalformed/utility"
       ].map { |f| Pathname.new(f) },
+      arduino_library_src_dirs: [],
       test_files: []
     },
-    "OnePointFiveDummy": {
+    OnePointFiveDummy: {
       one_five: true,
       cpp_files: [
         "OnePointFiveDummy/src/YesSrc.cpp",
         "OnePointFiveDummy/src/subdir/YesSubdir.cpp",
       ].map { |f| Pathname.new(f) },
+      cpp_files_libraries: [],
       header_dirs: [
         "OnePointFiveDummy/src",
         "OnePointFiveDummy/src/subdir",
       ].map { |f| Pathname.new(f) },
+      arduino_library_src_dirs: [],
       test_files: []
     }
-  }.freeze
+  }
+
+  # easier to construct this one from the other test cases
+  answers[:DependOnSomething] = {
+    one_five: true,
+    cpp_files: ["DependOnSomething/src/YesDeps.cpp"].map { |f| Pathname.new(f) },
+    cpp_files_libraries: answers[:OnePointOhDummy][:cpp_files] + answers[:OnePointFiveDummy][:cpp_files],
+    header_dirs: ["DependOnSomething/src"].map { |f| Pathname.new(f) }, # this is not recursive!
+    arduino_library_src_dirs: answers[:OnePointOhDummy][:header_dirs] + answers[:OnePointFiveDummy][:header_dirs],
+    test_files: []
+  }
+
+  answers.freeze
 
   answers.each do |sampleproject, expected|
     context "#{sampleproject}" do
       cpp_lib_path = sampleproj_path + sampleproject.to_s
-      cpp_library = ArduinoCI::CppLibrary.new(cpp_lib_path, Pathname.new("my_fake_arduino_lib_dir"), [])
+      cpp_library = ArduinoCI::CppLibrary.new(cpp_lib_path, sampleproj_path, [])
+      dependencies = cpp_library.arduino_library_dependencies.nil? ? [] : cpp_library.arduino_library_dependencies
 
       it "detects 1.5 format" do
         expect(cpp_library.one_point_five?).to eq(expected[:one_five])
@@ -75,6 +96,13 @@ RSpec.describe ArduinoCI::CppLibrary do
         it "finds cpp files in directory" do
           relative_paths = cpp_library.cpp_files.map { |f| get_relative_dir(f) }
           expect(relative_paths.map(&:to_s)).to match_array(expected[:cpp_files].map(&:to_s))
+        end
+      end
+
+      context "cpp_files_libraries" do
+        it "finds cpp files in directories of dependencies" do
+          relative_paths = cpp_library.cpp_files_libraries(dependencies).map { |f| get_relative_dir(f) }
+          expect(relative_paths.map(&:to_s)).to match_array(expected[:cpp_files_libraries].map(&:to_s))
         end
       end
 
@@ -99,6 +127,14 @@ RSpec.describe ArduinoCI::CppLibrary do
         it "finds cpp files in directory" do
           relative_paths = cpp_library.test_files.map { |f| get_relative_dir(f) }
           expect(relative_paths.map(&:to_s)).to match_array(expected[:test_files].map(&:to_s))
+        end
+      end
+
+      context "arduino_library_src_dirs" do
+        it "finds src dirs from dependent libraries" do
+          # we explicitly feed in the internal dependencies
+          relative_paths = cpp_library.arduino_library_src_dirs(dependencies).map { |f| get_relative_dir(f) }
+          expect(relative_paths.map(&:to_s)).to match_array(expected[:arduino_library_src_dirs].map(&:to_s))
         end
       end
     end